# server/docs/swagger.yaml
host: 127.0.0.1:1323
info:
  contact:
    name: API Support
  description: WISMED Wisite-API Server. Updated@ 2022-09-03T11:33:31+10:00
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: WISMED WISITE API
  version: "1.0"
paths:
  /api/admin/activate:
    put:
      consumes:
      - multipart/form-data
      parameters:
      - description: unique user name
        in: formData
        name: uname
        required: true
        type: string
      - description: 'true: activate, false: deactivate'
        in: formData
        name: flag
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK - action successfully
        "400":
          description: Fail - invalid true/false flag
        "401":
          description: Fail - unauthorized error
        "500":
          description: Fail - internal error
      security:
      - ApiKeyAuth: []
      summary: activate or deactivate a user
      tags:
      - Admin
  /api/admin/avatar:
    get:
      consumes:
      - application/json
      parameters:
      - description: uname for its avatar
        in: query
        name: uname
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK - get avatar src base64
        "400":
          description: Fail - user does not exist
        "404":
          description: Fail - avatar is empty
      security:
      - ApiKeyAuth: []
      summary: get a user's avatar src as base64
      tags:
      - Admin
  /api/admin/officialize:
    put:
      consumes:
      - multipart/form-data
      parameters:
      - description: unique user name
        in: formData
        name: uname
        required: true
        type: string
      - description: 'true: officialize, false: un-officialize'
        in: formData
        name: flag
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK - action successfully
        "400":
          description: Fail - invalid true/false flag
        "401":
          description: Fail - unauthorized error
        "500":
          description: Fail - internal error
      security:
      - ApiKeyAuth: []
      summary: officialize or un-officialize a user
      tags:
      - Admin
  /api/admin/onlines:
    get:
      consumes:
      - application/json
      parameters:
      - description: user filter with uname wildcard(*)
        in: query
        name: uname
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK - list successfully
        "401":
          description: Fail - unauthorized error
        "500":
          description: Fail - internal error
      security:
      - ApiKeyAuth: []
      summary: get all online users
      tags:
      - Admin
  /api/admin/spa/menu:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK - get menu successfully
        "500":
          description: Fail - internal error
      security:
      - ApiKeyAuth: []
      summary: get tailored side menu for different user group
      tags:
      - Admin
  /api/admin/users:
    get:
      consumes:
      - application/json
      parameters:
      - description: user filter with uname wildcard(*)
        in: query
        name: uname
        type: string
      - description: user filter with name wildcard(*)
        in: query
        name: name
        type: string
      - description: user filter with active status
        in: query
        name: active
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK - list successfully
        "401":
          description: Fail - unauthorized error
        "500":
          description: Fail - internal error
      security:
      - ApiKeyAuth: []
      summary: get all users' info
      tags:
      - Admin
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: authorization
    type: apiKey
swagger: "2.0"